t Adapter_SivisoListView delete_0_0databaseDelete

+++ AndroidStudio/app/src/main/java/com/jordan/betcher/siviso/siviso/list/Adapter_SivisoListView.java
@@ -33,6 +33,7 @@ extends RecyclerView.Adapter<ViewHolder_Abstract>
 	
 	public void deleteItem(int index)
 	{
+		if(index == 0) return;
 		database.delete(index-1);
 		//TODO
 		/*
